requirements.txt
fastapi~=0.112.2
uvicorn[standard]
sqlalchemy~=2.0.33
psycopg2-binary
pydantic~=2.8.2
python-dotenv~=0.10.5
passlib~=1.7.4
python-multipart
python-jose~=3.3.0
cryptography
bcrypt